+++ packages/form-data-parser/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This is the changelog for [`form-data-parser`](https://github.com/mjackson/remix-the-web/tree/main/packages/form-data-parser). It follows [semantic versioning](https://semver.org/).
 
+## HEAD
+
+- Export `FormDataParserError` and `MaxFilesExceededError`
+- Re-export `MultipartParseError`, `MaxHeaderSizeExceededError`, and `MaxFileSizeExceededError` from multipart parser
+
 ## v0.9.0 (2025-06-13)
 
 This release updates to `multipart-parser` 0.10.0 and removes the restrictions on checking the `size` and/or `slice`ing `FileUpload` objects.